Move geometry struct to geometry file
